Move collection show blade view into inertia ...
The View.vue was basically already the page component, but the props from the blade view are moved to the server.

This can obviously be cleaned up, but I converted it with as few changes as I could.
