+++ src/Http/Controllers/CP/Collections/CollectionsController.php
@@ -145,22 +145,51 @@ class CollectionsController extends CpController
             'actions' => Action::for($collection, ['view' => 'form']),
         ];
 
-        return Inertia::render('collections/Show');
-
         if ($collection->queryEntries()->count() === 0) {
             return view('statamic::collections.empty', $viewData);
         }
 
-        if (! $collection->hasStructure()) {
-            return view('statamic::collections.show', $viewData);
+        $user = \Statamic\Facades\User::current();
+        $props = [
+            'title' => $collection->title(),
+            'handle' => $collection->handle(),
+            'icon' => $collection->icon(),
+            'canCreate' => $viewData['canCreate'],
+            'createUrls' => $viewData['createUrls'],
+            'createLabel' => $collection->createLabel(),
+            'blueprints' => $blueprints->all(),
+            'sortColumn' => $collection->sortField(),
+            'sortDirection' => $collection->sortDirection(),
+            'columns' => $columns,
+            'filters' => $viewData['filters'],
+            'actions' => $viewData['actions'],
+            'actionUrl' => cp_route('collections.actions.run'),
+            'entriesActionUrl' => cp_route('collections.entries.actions.run', $collection->handle()),
+            'reorderUrl' => cp_route('collections.entries.reorder', $collection->handle()),
+            'editUrl' => $collection->editUrl(),
+            'blueprintsUrl' => cp_route('blueprints.collections.index', $collection),
+            'scaffoldUrl' => cp_route('collections.scaffold', $collection->handle()),
+            'canEdit' => $user->can('edit', $collection),
+            'canEditBlueprints' => $user->can('configure fields'),
+            'initialSite' => $site->handle(),
+            'sites' => $viewData['sites'],
+            'canChangeLocalizationDeleteBehavior' => $viewData['canChangeLocalizationDeleteBehavior'],
+        ];
+
+        if ($collection->hasStructure()) {
+            $structure = $collection->structure();
+            $props = [
+                ...$props,
+                'structured' => $user->can('reorder', $collection),
+                'structurePagesUrl' => cp_route('collections.tree.index', $structure->handle()),
+                'structureSubmitUrl' => cp_route('collections.tree.update', $collection->handle()),
+                'structureMaxDepth' => $structure->maxDepth() ?? PHP_FLOAT_MAX, // "Infinity"
+                'structureExpectsRoot' => $structure->expectsRoot(),
+                'structureShowSlugs' => $structure->showSlugs(),
+            ];
         }
 
-        $structure = $collection->structure();
-
-        return view('statamic::collections.show', array_merge($viewData, [
-            'structure' => $structure,
-            'expectsRoot' => $structure->expectsRoot(),
-        ]));
+        return Inertia::render('collections/Show', $props);
     }
 
     public function create()
